pyproject.toml
[tool.poetry]
name = "resolve-dependencies"
version = "0.1.0"
description = ""
authors = ["Mark Wallsgrove <markwallsgrove@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
poetry-core = "^2.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
